# specs/001-phase2-init/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo App API
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /api/{user_id}/tasks:
    get:
      summary: List Tasks
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed]
        - in: query
          name: sort
          schema:
            type: string
            enum: [asc, desc]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
    post:
      summary: Create Task
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get Task
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Not Found
    put:
      summary: Update Task
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
    delete:
      summary: Delete Task
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle Complete
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
    TaskRead:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time